% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_liab.R
\name{FGRS}
\alias{FGRS}
\title{FGRS}
\usage{
FGRS(
  df,
  stat_cols = which(grepl("stat", colnames(df))),
  age_cols = which(grepl("age", colnames(df))),
  rel_mat,
  prev,
  env_cor,
  aoo,
  ped
)
}
\arguments{
\item{df}{data.frame TODO}

\item{stat_cols}{TODO}

\item{rel_mat}{TODO}

\item{prev}{TODO}

\item{env_cor}{TODO}

\item{aoo}{TODO}

\item{ped}{TODO}

\item{age_cold}{TODO}
}
\description{
Our implementations of Kendler et al.
}
