% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_functions.R
\name{fgrs_accuracy}
\alias{fgrs_accuracy}
\title{FGRS accuracy}
\usage{
fgrs_accuracy(
  nrel = NULL,
  prev,
  h2,
  rel_rr = NULL,
  rel_ir = NULL,
  rel_matrix = NULL,
  ped = NULL,
  Nsim = 10000,
  method = c("pa", "linear", "linear2", "quantitative"),
  estimate = c("theory", "simulate"),
  ...
)
}
\arguments{
\item{nrel}{an optional integer denoting the number of relatives}

\item{prev}{numeric denoting the prevalence.}

\item{h2}{numeric denoting the heritability.}

\item{rel_rr}{if nrel is provided, the relationship coefficient between the relatives}

\item{rel_ir}{if nrel is provided, the relationship coefficient between the index person and the relatives}

\item{rel_matrix}{an optional (n x n) relatedness matrix}

\item{ped}{an optional object of class pedigree}

\item{Nsim}{if estimate = "simulate", the number of families to be simulated.}

\item{method}{the method for estimating the accuracy. Can be "pa" for the pa-fgrs method, "linear" or "linear2" for a linear predictor or "quantitative" for a linear predictor of a quantitative trait.}

\item{estimate}{the method for generating the estimate. Can be "theory" for the theoretical expectation or "simulate" for runnning a simulation.}

\item{...}{additional params}
}
\description{
Estimate the expected accuracy of a family genetic risk score given a configuration of family members and the prevalence and heritability of the phenotype
The input can have one of three formats:
it can be a pedigree object generated by kinship2::pedigree(),
it can a (n x n) relatedness matrix or if all relatives have the same relationship to the index person and to each other,
the imput can be the number of relatives, the relatedness coefficient between the index person and the relatives, and the
}
\examples{
fgrs_accuracy(nrel=2, rel_rr=0.5, rel_ir=0.5, prev=0.1, h2=0.5, method= "pa", estimate="theory")
}
