% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{FGRS_wrapper}
\alias{FGRS_wrapper}
\title{FGRS wrapper}
\usage{
FGRS_wrapper(
  proband_ids,
  K,
  pheno,
  method = "PA",
  thr = NULL,
  w = NULL,
  h2 = NULL,
  env_cor_s = 1,
  env_cor_f = 1,
  env_cor_m = 1,
  sib_mat = NULL,
  father_mat = NULL,
  mother_mat = NULL
)
}
\arguments{
\item{proband_ids}{vector of sample ids indicating which individuals FGRS should be estimated for.}

\item{K}{kinship matrix provided either as a \code{matrix}, a \code{dsCMatrix} or a \code{data.frame} with column names \code{i}, \code{j} and \code{x}.}

\item{pheno}{\code{data.frame} containing the phenotype information on the relatives. Must contain columns called \code{id} and \code{aff}.}

\item{method}{character indicating which FGRS method to use. Default is \code{method="PA"}.}

\item{thr}{numeric vector wiht threshold value used for each relative in \code{pheno}.}

\item{w}{numeric vector of proportion of risk experienced by each relative in \code{pheno}.}

\item{h2}{numeric heritability estimate of the phenotype (liability scale).}

\item{env_cor_s}{numeric indicating the down-weighting of siblings (only used for Ohlsson-Kendler FGRS). Default =1 (no correction).}

\item{env_cor_f}{numeric indicating the down.weighting of fathers (only used for Ohlsson-Kendler FGRS). Default =1 (no correction).}

\item{env_cor_m}{numeric indicating the down-weighting of mothers (only used for Ohlsson-Kendler FGRS). Default =1 (no correction).}

\item{sib_mat}{optional \code{data.frame}  with colunm names \code{i} and \code{j} giving the \code{id}s of individuals that are siblings.}

\item{father_mat}{optional \code{data.frame}  with colunm names \code{i} and \code{j} giving the \code{id}s of individuals that are father-child relations.}

\item{mother_mat}{optional \code{data.frame}  with colunm names \code{i} and \code{j} giving the \code{id}s of individuals that are mother-child relations.}
}
\value{
postM posterior mean liability

postVar posterior variance
}
\description{
Compute (PA-)FGRS from for a large a dataset
}
\examples{
pa_fgrs(c(0,1),qnorm(.9),covmat = matrix(c(.5,.25,.25,.25,1,.25,.25,.25,1),3))
}
