% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_sparse.R
\name{kinship_path}
\alias{kinship_path}
\title{kinship sparse path}
\usage{
kinship_path(ped)
}
\arguments{
\item{pedigree}{}
}
\description{
Compute a sparse kinship matrix from a pedigree object using path counting
}
\details{
This is an alternative way of computing the kinhips matrix which is usually slower, but less memory intense.
}
\examples{
library(kinship2)
data(minnbreast)
system.time(k_path <- kinship_path(with(minnbreast[minnbreast$famid \%in\% 4:5,],pedigree(id, fatherid, motherid, sex))))
system.time(k_sparse <- kinship_sparse(with(minnbreast[minnbreast$famid \%in\% 4:5,],pedigree(id, fatherid, motherid, sex))))
system.time(k_per_fam <- kinship(with(minnbreast[minnbreast$famid \%in\% 4:5,],pedigree(id, fatherid, motherid, sex, famid=famid))))
identical(k_path,k_sparse) 
sum(!k_path[order(rownames(k_path)),order(rownames(k_path))]==k_per_fam[order(rownames(k_per_fam)),order(rownames(k_per_fam))]) 
}
