% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\docType{data}
\name{minnbreast}
\alias{minnbreast}
\title{Minnbreast Data from kinship2}
\format{
A data frame with 28081 rows and 15 columns.
}
\source{
\pkg{kinship2} package
}
\usage{
data(minnbreast)
}
\description{
A brief description of the \code{minnbreast} dataset.
Provide context, source, and any other relevant details.
}
\details{
Detailed information about the dataset, if available.
This can include variable descriptions, data collection methods, etc.
}
\examples{
data(minnbreast)
head(minnbreast)


}
