% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_liab.R
\name{pa_fgrs}
\alias{pa_fgrs}
\title{PA-FGRS}
\usage{
pa_fgrs(
  rel_status,
  thr = NA,
  rel_thr = rep(thr, length(rel_status)),
  rel_w = rep(1, length(rel_status)),
  covmat,
  i_status = NA,
  i_thr = thr,
  i_w = 1,
  conditional.mix = T
)
}
\arguments{
\item{rel_status}{vector with status of relatives 1=affected, 0=not affected, NA= missing}

\item{thr}{threshold value used for all individuals (relatives and index person)}

\item{rel_thr}{optional vector of specific threshold for each relative}

\item{rel_w}{optional vector of proportion of risk experienced by each relative}

\item{covmat}{covariance matrix of liabilities. First row and column correspoding to index person,subsequent rows and columns corresponding to relatives.}

\item{i_status}{optional status of index person, if to be considered 1=affected, 0=not affected, NA= missing}

\item{i_thr}{optional specific threshold for index person}

\item{i_w}{optional proportion of risk experienced by index person}
}
\description{
Compute posterior mean and variance of liability of index person given disease status, thresholds and proportion of risk observed in relatives.
}
\examples{
pa_fgrs(c(0,1),qnorm(.9),covmat = matrix(c(.5,.25,.25,.25,1,.25,.25,.25,1),3))

}
