% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converters.R
\name{ped2graph}
\alias{ped2graph}
\title{Convert Pedigree Data to Graph}
\usage{
ped2graph(ped)
}
\arguments{
\item{ped}{A data frame or an object of class "pedigree" or "pedigreeList" containing pedigree data.}
}
\value{
An \code{igraph} graph object representing the pedigree data.
}
\description{
This function converts a pedigree data frame to a graph representation using the \code{igraph} package.
It was inspired by a similar function in the \code{FamAgg} package.
}
\details{
It first checks if the input is of class "pedigree" or "pedigreeList" and converts it to a data frame.
It then ensures the pedigree data has the required columns and creates a graph representation using the \code{igraph} package.
}
\examples{
# Create a sample pedigree data frame
ped_data <- data.frame(
  family = c(1, 1, 1),
  id = c(1, 2, 3),
  father = c(NA, NA, 1),
  mother = c(NA, NA, 2),
  sex = c("M", "F", "M")
)

# Convert the pedigree data to a graph
graph <- ped2graph(ped_data)

# Plot the graph
plot(graph)

}
\seealso{
\url{https://github.com/EuracBiomedicalResearch/FamAgg} for the original \code{FamAgg} package.
}
