% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_liab.R
\name{FGRS}
\alias{FGRS}
\title{FGRS}
\usage{
FGRS(
  df,
  stat_cols = which(grepl("stat", colnames(df))),
  age_cols = which(grepl("age", colnames(df))),
  rel_mat,
  prev,
  env_cor = 1,
  aoo
)
}
\arguments{
\item{df}{data.frame with one row per individual with columns containing the phenotype information on the relatives}

\item{stat_cols}{vector of names of the columns containing the disease status of the relatives}

\item{age_cols}{vector of names of the columns containing the age of the relatives}

\item{rel_mat}{relatedness matrix with the first row and column correponding to the index person a subsequent rows to the relatives.}

\item{prev}{numeric indicating disease prevalence}

\item{env_cor}{optional numeric. Factor by which the first degree relatives are downweighted.}

\item{aoo}{a function that transforms the age values into a proportion of risk experienced by the individual}
}
\description{
Our implementations of Kendler et al.
}
