% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship_sparse.R
\name{kinship_path}
\alias{kinship_path}
\title{kinship sparse path}
\usage{
kinship_path(ped)
}
\arguments{
\item{ped}{a pedigree object}
}
\description{
Compute a sparse kinship matrix from a pedigree object using path counting
}
\details{
This is an alternative way of computing the kinhips matrix which is usually slower, but less memory intense.
}
\examples{
library(kinship2)
data(minnbreast)
mb <- minnbreast[minnbreast$famid \%in\% 4:5,]
system.time(k_path <- kinship_path(
     pedigree(mb$id, mb$fatherid, mb$motherid, mb$sex)))
system.time(k_sparse <- kinship_sparse(
     pedigree(mb$id, mb$fatherid, mb$motherid, mb$sex)))
system.time(k_per_fam <- kinship(
     pedigree(mb$id, mb$fatherid, mb$motherid, mb$sex, famid=mb$famid)))
identical(k_path,k_sparse) 
val1=!k_path[order(rownames(k_path)),order(rownames(k_path))]
val2=k_per_fam[order(rownames(k_per_fam)),order(rownames(k_per_fam))]
sum(val1==val2)
}
